Implement usePeer hook for remote control
