+++ src/pages/Document.tsx
@@ -1,74 +1,23 @@
-import React, { useState, useEffect, useRef, useMemo } from 'react';
+import React, { useState, useEffect, useRef } from 'react';
 import { useParams } from 'react-router-dom';
-
-import { DottingDoc } from '../types/document.js';
-import { initialDataArray, initialIndices } from '../data/initialData.js';
-import { activateClient, attachDocument, createDocument } from '../utils/document.js';
-
-import yorkie, { DocEvent, Indexable, JSONArray } from 'yorkie-js-sdk';
+import { DocEvent } from 'yorkie-js-sdk';
 import { Dotting, useData, useDotting, useHandlers, DottingRef, PixelModifyItem } from 'dotting';
 
+import usePeer from '../hooks/usePeer.js';
+import { initialDataArray } from '../data/initialData.js';
+
 export default function Document() {
-  /* DOCUMENT ID */
+  /* Document Id */
   const { docId } = useParams<{ docId: string }>();
 
-  /* DOTTING */
+  /* Dotting */
   const ref = useRef<DottingRef>(null);
   const { dataArray } = useData(ref);
   const { colorPixels, setData } = useDotting(ref);
   const { addStrokeEndListener, removeStrokeEndListener } = useHandlers(ref);
 
-  /* YORKIE */
-  const [doc, setDoc] = useState<yorkie.Document<DottingDoc>>();
-  const [client, setClient] = useState<yorkie.Client<Indexable>>();
-  const [isMultiplayerReady, setIsMultiplayerReady] = useState<boolean>(false);
-
-  /* LOCAL STATES */
-  const isDataLoaded = useMemo(() => {
-    return dataArray.length !== 0;
-  }, [dataArray]);
-
-  const initializeRemoteData = () => {
-    doc?.update((root) => {
-      if (!root.indices) {
-        root.indices = initialIndices;
-      }
-
-      if (!root.data) {
-        root.data = {};
-        dataArray.forEach((row) => {
-          row.forEach(({ rowIndex, columnIndex, color }) => {
-            if (!root.data[rowIndex]) {
-              root.data[rowIndex] = {};
-            }
-            if (!root.data[rowIndex][columnIndex]) {
-              // @ts-ignore
-              root.data[rowIndex][columnIndex] = {};
-            }
-            root.data[rowIndex][columnIndex].color = color;
-            root.data[rowIndex][columnIndex].rowIndex = rowIndex;
-            root.data[rowIndex][columnIndex].columnIndex = columnIndex;
-          });
-        });
-      } else {
-        const pixelArray: Array<Array<PixelModifyItem>> = [];
-        const rawRecord: Record<string, Record<string, PixelModifyItem>> = JSON.parse(
-          // @ts-ignore
-          root.data.toJSON(),
-        );
-        Object.values(rawRecord).forEach((record) => {
-          pixelArray.push([]);
-          Object.values(record).forEach((item) => {
-            pixelArray[pixelArray.length - 1].push(item);
-          });
-        });
-        setData(pixelArray);
-      }
-    }, 'Initialize default Dotting Document fields if not exists');
-
-    client.sync();
-    setIsMultiplayerReady(true);
-  };
+  /* Get data from hook */
+  const { doc, client, isMultiplayerReady } = usePeer({ docId, dataArray, setData });
 
   useEffect(() => {
     const updateData = ({ strokedPixels }) => {
@@ -83,7 +32,7 @@ export default function Document() {
     return () => {
       removeStrokeEndListener(updateData);
     };
-  }, [addStrokeEndListener, removeStrokeEndListener, doc]);
+  }, [doc, addStrokeEndListener, removeStrokeEndListener]);
 
   useEffect(() => {
     if (!isMultiplayerReady) return;
@@ -107,28 +56,7 @@ export default function Document() {
         colorPixels(pixelsToColor);
       }
     });
-  }, [isMultiplayerReady, client, doc]);
-
-  useEffect(() => {
-    activateClient().then((client) => {
-      setClient(client);
-    });
-  }, [docId]);
-
-  useEffect(() => {
-    if (!client) return;
-    const doc = createDocument(docId);
-    setDoc(doc);
-  }, [client]);
-
-  useEffect(() => {
-    if (!client || !doc) return;
-    if (!isDataLoaded) return;
-
-    attachDocument(client, doc).then(() => {
-      initializeRemoteData();
-    });
-  }, [client, doc, isDataLoaded]);
+  }, [doc, client, isMultiplayerReady]);
 
   return (
     <>
